Implement "Void"
Simple change to merge Item and Liquid Void.
