+++ core/assets-raw/sprites/blocks/production/void.png
