+++ core/assets/bundles/bundle.properties
@@ -1020,9 +1020,8 @@ block.cultivator.name = Cultivator
 block.conduit.name = Conduit
 block.mechanical-pump.name = Mechanical Pump
 block.item-source.name = Item Source
-block.item-void.name = Item Void
+block.void.name = Void
 block.liquid-source.name = Liquid Source
-block.liquid-void.name = Liquid Void
 block.power-void.name = Power Void
 block.power-source.name = Power Infinite
 block.unloader.name = Unloader
