+++ core/src/mindustry/world/blocks/sandbox/ItemVoid.java
@@ -9,6 +9,9 @@ public class ItemVoid extends Block{
     public ItemVoid(String name){
         super(name);
         update = solid = acceptsItems = true;
+        hasLiquids = true;
+        solid = true;
+        update = true;
     }
 
     public class ItemVoidEntity extends Building{
@@ -18,6 +21,14 @@ public class ItemVoid extends Block{
         @Override
         public boolean acceptItem(Building source, Item item){
             return true;
+        }
+        @Override
+        public boolean acceptLiquid(Building source, Liquid liquid, float amount){
+            return true;
+        }
+
+        @Override
+        public void handleLiquid(Building source, Liquid liquid, float amount){
         }
     }
 }
